pyproject.toml
[project]
name = 'trmnl'
version = '0.0.1'
requires-python = '==3.13.*'
dependencies = [
    "pillow ==11.1.0",
]

[tool.uv]
dev-dependencies = [
    "mypy ==1.15.0",
    "pytest ==8.3.4",
    "pytest-cov ==6.0.0",
]

